add additional condition to store, update & delete method to check permission for group session user

+++ app/Http/Controllers/RecordController.php
@@ -8,7 +8,9 @@ use App\Models\Account;
 use App\Models\Category;
 use Illuminate\Http\Request;
 
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
+use Spatie\Permission\Models\Permission;
 use Illuminate\Support\Facades\Validator;
 
 class RecordController extends Controller
@@ -147,7 +149,6 @@ class RecordController extends Controller
         return view('record.create', compact('accounts', 'categories'));
     }
 
-
     /**
      *  Store record
      */
@@ -168,18 +169,34 @@ class RecordController extends Controller
 
         if ($validator->fails()) {
             return back()->withErrors($validator)->withInput();
-        } else {
-            $record = new Record;
-            $record->account_id = $request->input('account_id');
-            $record->category_id = $request->input('category_id');
-            $record->user_id = auth()->id();
-            $record->type = $request->input('type');
-            $record->amount = $request->input('amount');
-            $record->datetime = $request->input('datetime');
-            $record->description = $request->input('description');
-            $record->expense_sharing_group_id = $request->input('group_id');
-            $record->save();
         }
+
+        $userId = auth()->user()->id;
+        $groupId = $request->input('group_id');
+
+        if ($groupId != null) {
+            $groupMember = DB::table('group_members')->where('user_id', $userId)
+                ->where('expense_sharing_group_id', $groupId)
+                ->first();
+
+            $requiredPermission = Permission::where('name', 'create record')->first();
+            $permissionsArray = json_decode($groupMember->permissions, true);
+            if (!$permissionsArray || !in_array($requiredPermission->id, $permissionsArray)) {
+                return redirect()->route('record.index')->with('error', 'You do not have permission to create a record in this group.');
+            }
+        }
+
+        $record = new Record;
+        $record->account_id = $request->input('account_id');
+        $record->category_id = $request->input('category_id');
+        $record->user_id = auth()->id();
+        $record->type = $request->input('type');
+        $record->amount = $request->input('amount');
+        $record->datetime = $request->input('datetime');
+        $record->description = $request->input('description');
+        $record->expense_sharing_group_id = $request->input('group_id');
+        $record->save();
+
 
         return redirect()->route('record.index')->with('success', 'Record added successfully');
     }
@@ -223,8 +240,22 @@ class RecordController extends Controller
         );
 
         if ($validator->fails()) {
-            return redirect()->route('record.edit')
-                ->withErrors($validator);
+            return redirect()->route('record.edit')->withErrors($validator);
+        }
+
+        $userId = auth()->user()->id;
+        $groupId = $request->input('group_id');
+
+        if ($groupId != null) {
+            $groupMember = DB::table('group_members')->where('user_id', $userId)
+                ->where('expense_sharing_group_id', $groupId)
+                ->first();
+
+            $requiredPermission = Permission::where('name', 'edit record')->first();
+            $permissionsArray = json_decode($groupMember->permissions, true);
+            if (!$permissionsArray || !in_array($requiredPermission->id, $permissionsArray)) {
+                return redirect()->route('record.index')->with('error', 'You do not have permission to create a record in this group.');
+            }
         }
 
         $record->update([
@@ -246,10 +277,32 @@ class RecordController extends Controller
     public function delete(Record $record)
     {
         $userId = Auth::id();
-        if ($record->user_id === $userId) {
+        
+        if ($record->user_id === $userId && is_null($record->expense_sharing_group_id)) {
             $record->delete();
+            return redirect()->route('record.index')->with('success', 'Record deleted successfully');
         }
 
-        return redirect()->route('record.index')->with('success', 'Record deleted successfully');
+        // If it's not the personal budget, check for group permission
+        $activeGroupId = session('active_group_id');
+        $groupMember = DB::table('group_members')->where('user_id', $userId)
+            ->where('expense_sharing_group_id', $activeGroupId)
+            ->first();
+
+        if ($groupMember) {
+            // Check if the user has the necessary permission to delete budgets within the group
+            $requiredPermission = Permission::where('name', 'delete record')->first();
+
+            $permissionsArray = json_decode($groupMember->permissions, true);
+
+            if ($permissionsArray && in_array($requiredPermission->id, $permissionsArray)) {
+                // The user has permission, delete the budget
+                $record->delete();
+                return redirect()->route('record.index')->with('success', 'Record deleted successfully');
+            }
+        }
+
+        // Permission denied
+        return redirect()->route('record.index')->with('error', 'You do not have permission to delete this record.');
     }
 }
